% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermanPlot.R
\name{GermanPlot}
\alias{GermanPlot}
\title{create a German Map}
\description{
create a German Map

create a German Map
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GermanPlot-new}{\code{GermanPlot$new()}}
\item \href{#method-GermanPlot-osm}{\code{GermanPlot$osm()}}
\item \href{#method-GermanPlot-join_features}{\code{GermanPlot$join_features()}}
\item \href{#method-GermanPlot-plot_feature}{\code{GermanPlot$plot_feature()}}
\item \href{#method-GermanPlot-clone}{\code{GermanPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermanPlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-GermanPlot-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermanPlot$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
R6
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermanPlot-osm"></a>}}
\if{latex}{\out{\hypertarget{method-GermanPlot-osm}{}}}
\subsection{Method \code{osm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermanPlot$osm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermanPlot-join_features"></a>}}
\if{latex}{\out{\hypertarget{method-GermanPlot-join_features}{}}}
\subsection{Method \code{join_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermanPlot$join_features(data, ebene = "kreis")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermanPlot-plot_feature"></a>}}
\if{latex}{\out{\hypertarget{method-GermanPlot-plot_feature}{}}}
\subsection{Method \code{plot_feature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermanPlot$plot_feature(feature, alpha = 0.8, ebene = "kreis")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermanPlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GermanPlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermanPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
