% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_functions.R
\name{morans_i}
\alias{morans_i}
\title{Calculate Moran's I}
\usage{
morans_i(geometry, var, queen = T, summary = T)
}
\arguments{
\item{geometry}{data with geometry column that includes spatial information}

\item{var}{vector for which Moran's I should be calculated, should be in the same order as geometry}

\item{queen}{detect neighbors based on shared points yes or no}

\item{summary}{only report the most import estimates}
}
\description{
Calculate Moran's I
}
