% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_functions.R
\name{morans_plot}
\alias{morans_plot}
\title{Scatter Plot that compares lagged means of neighbors with values in a region}
\usage{
morans_plot(geometry, var, queen = T)
}
\arguments{
\item{geometry}{data with geometry column that includes spatial information}

\item{var}{vector for which Moran's I should be calculated, should be in the same order as geometry}

\item{queen}{detect neighbors based on shared points yes or no}
}
\description{
Scatter Plot that compares lagged means of neighbors with values in a region
}
